refactor(MenuManagers): Moved to appropriate namespace/folder.
