+++ Assets/UnZENity-Core/Scripts/UI/IMenu.cs
@@ -1,0 +1,7 @@
+﻿namespace GUZ.Core.UI
+{
+    public interface IMenu
+    {
+        public void ToggleVisibility();
+    }
+}
