+++ Assets/UnZENity-Core/Scripts/UI/Main/MovementMenuHandler.cs
@@ -3,7 +3,7 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace GUZ.Core.UI.MainMenu
+namespace GUZ.Core.UI.Main
 {
     public class MovementMenuHandler : MonoBehaviour
     {
