+++ Assets/UnZENity-Core/Scripts/UI/StatusMenu.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using UnityEngine;
+
+namespace GUZ.Core.UI
+{
+    public class StatusMenu : MonoBehaviour, IMenu
+    {
+        public void ToggleVisibility()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
